a lot of experiments with lights shadows etc, also create render passes and move logic from render to renderSystem, now which pass is separate logic object which controlled by renderSystem

+++ src/ecsModule/SystemManager.h
@@ -71,7 +71,7 @@ namespace ecsModule {
 		}
 
 		template <class T>
-		void setSystemPriority(size_t newPriority) {
+		void setSystemPriority(uint16_t newPriority) {
 			if (auto system = getSystem<T>()) {
 				if (system->mPriority == newPriority) {
 					return;
